Feature/error

+++ src/main/java/com/example/memorip/service/TimelineService.java
@@ -3,15 +3,17 @@ package com.example.memorip.service;
 import com.example.memorip.dto.TimelineDTO;
 import com.example.memorip.entity.Plan;
 import com.example.memorip.entity.Timeline;
+import com.example.memorip.exception.CustomException;
+import com.example.memorip.exception.ErrorCode;
 import com.example.memorip.repository.PlanRepository;
 import com.example.memorip.repository.TimelineMapper;
 import com.example.memorip.repository.TimelineRepository;
-import com.example.memorip.exception.NotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class TimelineService {
@@ -28,7 +30,7 @@ public class TimelineService {
     public Timeline save(TimelineDTO timelineDTO){
         int planId = timelineDTO.getPlanId();
         Plan plan = planRepository.findById(planId);
-        if(plan==null) throw new NotFoundException("Plan not found with id: " + planId);
+        if(plan==null) throw new CustomException(ErrorCode.PLAN_NOT_FOUND);
 
         Timeline timeline = TimelineMapper.INSTANCE.timelineDTOToTimeline(timelineDTO);
         timeline.setPlan(plan);
@@ -40,19 +42,23 @@ public class TimelineService {
     @Transactional(readOnly = true)
     public List<Timeline> findByPlanId(int planId){
         Plan plan = planRepository.findById(planId);
-        if(plan==null) throw new NotFoundException("Plan not found with id: " + planId);
+        if(plan==null) throw new CustomException(ErrorCode.PLAN_NOT_FOUND);
 
         return timelineRepository.findAllByPlanId(planId);
     }
 
     @Transactional(readOnly = true)
     public Timeline findOneById(int id){
-        return timelineRepository.findById(id).orElseThrow(()->new NotFoundException("Timeline not found with id: " + id));
+        return timelineRepository.findById(id).orElseThrow(()->new CustomException(ErrorCode.TIMELINE_NOT_FOUND));
     }
 
     @Transactional
     public void deleteById(int id){
-        timelineRepository.findById(id).orElseThrow(()->new NotFoundException("Timeline not found with id: " + id));
+        Optional<Timeline> timeline = timelineRepository.findById(id);
+        if(timeline.isEmpty()){
+            throw new CustomException(ErrorCode.TIMELINE_NOT_FOUND);
+        }
         timelineRepository.deleteById(id);
     }
+
 }
